fix: change time_limit type from float to int in problem_update_info
Polygon API requires timeLimit to be an integer. The float type annotation
caused FastMCP to generate a `number` JSON schema, leading to values being
rejected with "Field should contain integer value".

+++ polygon_mcp/server.py
@@ -430,7 +430,7 @@ def problem_update_info(
     input_file: Optional[str] = None,
     output_file: Optional[str] = None,
     interactive: Optional[bool] = None,
-    time_limit: Optional[float] = None,
+    time_limit: Optional[int] = None,
     memory_limit: Optional[int] = None,
 ) -> Any:
     polygon = _get_client()
